niffler-e-2-e-tests-python/requirements.txt
pytest~=8.3.2
faker~=28.4.1
requests~=2.32.3
python-dotenv~=1.0.1
allure-pytest
requests-toolbelt

psycopg2-binary
sqlalchemy>=2.0.34
sqlmodel~=0.0.22
playwright~=1.46.0
pydantic~=2.9.0
pytest-playwright
python-dateutil~=2.9.0.post0
curlify